% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_filemaker.R
\docType{data}
\name{file}
\alias{file}
\title{Parse filemaker strains DB}
\format{An object of class \code{character} of length 1.}
\usage{
file
}
\arguments{
\item{file}{Path to filemaker CSV.}

\item{write}{Should the resulting tables be written. Defaults to \code{FALSE}.}
}
\description{
Parse the filemaker strains DB from an exported CSV.
}
\keyword{datasets}

