% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_filemaker.R
\name{parse_filemaker_export}
\alias{parse_filemaker_export}
\title{Parse filemaker strains DB}
\usage{
parse_filemaker_export(file, write = TRUE, blame = TRUE)
}
\arguments{
\item{file}{Path to filemaker CSV.}

\item{write}{Should the resulting tables be written. Defaults to \code{FALSE}.}
}
\description{
Parse the filemaker strains DB from an exported CSV.
}

