% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rothfreezer-package.R
\docType{package}
\name{rothfreezer}
\alias{rothfreezer}
\alias{rothfreezer-package}
\title{Rothfreezer R package}
\description{
Rothfreezer R package
}
\author{
Eric Bryant \email{eeb2139@columbia.edu},

Maintainer: Eric Bryant \email{eeb2139@columbia.edu}
}

