% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rothfreezer.R
\name{src_rothfreezer}
\alias{rothfreezer}
\alias{src_rothfreezer}
\title{Connect to the rothfreezer database}
\usage{
src_rothfreezer()
}
\description{
Connect to the rothfreezer SQLite database.
}
\examples{
# For a description of each table
vignette('database-overview', package = 'rothfreezer')

# Connect to the database
db <- src_rothfreezer()
db

}

