% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_from_googlesheet}
\alias{update_from_googlesheet}
\alias{update_antibodies}
\alias{update_primers}
\alias{update_media}
\alias{update_queries}
\alias{update_treatments}
\alias{update_strain_collections}
\alias{update_strain_collection_cispt_sensitive_1}
\alias{update_strain_collection_MATa_cispt_384_hb_V2}
\alias{update_strain_collection_MATa_cispt_384_hb}
\alias{update_strain_collection_MATa_cispt_96_hb_V2}
\alias{update_strain_collection_MATa_cispt_96_hb}
\alias{update_strain_collection_MATa_deletion_384}
\alias{update_strain_collection_MATalpha_deletion_384_hb}
\alias{update_strain_collection_MATalpha_deletion_96}
\alias{update_strain_collection_MATalpha_CTF4_RAD5_384_hb}
\alias{update_strain_collection_MATalpha_CTF4_RAD5_96_hb}
\alias{update_strain_collection_MATa_ts_384_hb}
\alias{update_strain_collection_info}
\alias{update_plasmids}
\alias{update_plasmids_pWJ}
\alias{update_plasmids_ZOO}
\alias{update_strains}
\alias{update_strains_special}
\alias{update_strains_J}
\alias{update_strains_R}
\alias{update_strains_rec}
\alias{update_strains_TS}
\alias{update_strains_U}
\alias{update_strains_W}
\title{Update a CSV file from a google spreadsheet}
\usage{
update_from_googlesheet(file, from)

update_antibodies(file = "antibodies/antibodies.csv",
  from = "https://docs.google.com/spreadsheets/d/1gyiZth6awGqdB6YiQqDLL9WSha7KxI525GYY6aoxcCI/edit?usp=sharing")

update_primers(file = "primers/primers.csv",
  from = "https://docs.google.com/spreadsheets/d/1tYxhHsEN4mBGJRZXqJS7vK7TJJdP7Wayv4P14R9wTMI/edit?usp=sharing")

update_media(file = "media/media.csv",
  from = "https://docs.google.com/spreadsheets/d/1mG2Vr1sMnNAIX-NCFC33O0WoUVDrI3F6no78Czs6Y8U/edit?usp=sharing")

update_queries(file = "queries/queries.csv",
  from = "https://docs.google.com/spreadsheets/d/1VgVEg1iJrbjOU6kh_9i4ZxmINBaVkrxoF-kGl0m51WM/edit?usp=sharing")

update_treatments(file = "treatments/treatments.csv",
  from = "https://docs.google.com/spreadsheets/d/1pWgFVl_rQ2HUtPSIthf0REjGvLNEUfzr7ob9rRAw2NI/edit?usp=sharing")

update_strain_collections()


  update_strain_collection_cispt_sensitive_1(file = "strain-collections/cispt-sensitive-1.csv",
  from = "https://docs.google.com/spreadsheets/d/1vTC2im5t2u-PW4QJNNia_M2fsAVXqodOj3UPmhJ7-4A/edit?usp=sharing")


  update_strain_collection_MATa_cispt_384_hb_V2(file = "strain-collections/MATa-cispt-384-hb-V2.csv",
  from = "https://docs.google.com/spreadsheets/d/1nHFLrPqMODgpJaBkkcDniEnp8Cq7tA8iOowvCPI5wuQ/edit?usp=sharing")


  update_strain_collection_MATa_cispt_384_hb(file = "strain-collections/MATa-cispt-384-hb.csv",
  from = "https://docs.google.com/spreadsheets/d/1r9VKumSwjGvEcBP8XebfCsXJnKDHvh_AnCKiiBty3rk/edit?usp=sharing")


  update_strain_collection_MATa_cispt_96_hb_V2(file = "strain-collections/MATa-cispt-96-hb-V2.csv",
  from = "https://docs.google.com/spreadsheets/d/1eCjWKnrF-H9A0qdaufYuir3CdIy-ueMffLv3MVezjp0/edit?usp=sharing")


  update_strain_collection_MATa_cispt_96_hb(file = "strain-collections/MATa-cispt-96-hb.csv",
  from = "https://docs.google.com/spreadsheets/d/1f9XuSwpJeN-j5_wQ3mwUGOlsGH-dixgBnHeamzl7PzU/edit?usp=sharing")


  update_strain_collection_MATa_deletion_384(file = "strain-collections/MATa-deletion-384.csv",
  from = "https://docs.google.com/spreadsheets/d/1dl89Nf4oI9c6lhAxAfApdM1-mz-RqkzZbrE4DOPQjxk/edit?usp=sharing")


  update_strain_collection_MATalpha_deletion_384_hb(file = "strain-collections/MATalpha-deletion-384-hb.csv",
  from = "https://docs.google.com/spreadsheets/d/1yyBZh8OEfjmR0oMENEqgCgUj03STMELHIZ08_Kl_fHQ/edit?usp=sharing")


  update_strain_collection_MATalpha_deletion_96(file = "strain-collections/MATalpha-deletion-96.csv",
  from = "https://docs.google.com/spreadsheets/d/1whG-KPW9htHvIR8LutivZpW3w3pEBSdEtJpg66aOaq0/edit?usp=sharing")


  update_strain_collection_MATalpha_CTF4_RAD5_384_hb(file = "strain-collections/MATalpha-CTF4,RAD5-384-hb.csv",
  from = "https://docs.google.com/spreadsheets/d/145J9KpYN38Ra8TAw2-dvAc_g80EJ_iAeWtqqp-NlPf8/edit?usp=sharing")


  update_strain_collection_MATalpha_CTF4_RAD5_96_hb(file = "strain-collections/MATalpha-CTF4,RAD5-96-hb.csv",
  from = "https://docs.google.com/spreadsheets/d/19CJrfjAgspW-zE2dkRa9nvQDsrfuWYO45dc3PSxiQII/edit?usp=sharing")


  update_strain_collection_MATa_ts_384_hb(file = "strain-collections/MATa-ts-384-hb.csv",
  from = "https://docs.google.com/spreadsheets/d/1zlM9wxoVHtSk5yac209veMCGZPvOnGRUdt5OJSqdGmA/edit?usp=sharing")


  update_strain_collection_info(file = "strain-collection-info/strain-collection-info.csv",
  from = "https://docs.google.com/spreadsheets/d/12osNdQV8Al6lVhvssnvnVajLXxMZ7qH6Bv8bwvrUg5c/edit?usp=sharing")

update_plasmids()

update_plasmids_pWJ(file = "plasmids/pWJ.csv",
  from = "https://docs.google.com/spreadsheets/d/15DXiQUBR4QQsthB7v9IU84ZcmiZB-kBKfrOjpiyisvw/edit?usp=sharing")

update_plasmids_ZOO(file = "plasmids/ZOO.csv",
  from = "https://docs.google.com/spreadsheets/d/1P-F7s-SGvey6B3_PWf-HGv07iwjkEqhwuz-muSyQuyU/edit?usp=sharing")

update_strains()

update_strains_special(file = "strains/special.csv",
  from = "https://docs.google.com/spreadsheets/d/1M-3LGm8P4X_fQ4hYO9BRha1JvwrmTHVExGRAwHXA_kw/edit?usp=sharing")

update_strains_J(file = "strains/J.csv",
  from = "https://docs.google.com/spreadsheets/d/1XaqVTIWVFpV4PWi2-N8qUrZJZpxuRfDUqdB8CdsmnMw/edit?usp=sharing")

update_strains_R(file = "strains/R.csv",
  from = "https://docs.google.com/spreadsheets/d/1UT1e4LyINK1F_z_koIpNC217R-om5fKL1IdwRPt0CLI/edit?usp=sharing")

update_strains_rec(file = "strains/rec.csv",
  from = "https://docs.google.com/spreadsheets/d/19UL9r7UNu72PAZ6fqXIJ8TpP1YB5JfCcpcknPCrw0W8/edit?usp=sharing")

update_strains_TS(file = "strains/TS.csv",
  from = "https://docs.google.com/spreadsheets/d/1J4MvBQTcJJLk2umR5jhsXPqn1SZ-SNWi763nBq7mNQY/edit?usp=sharing")

update_strains_U(file = "strains/U.csv",
  from = "https://docs.google.com/spreadsheets/d/1sj6cKonhGEcsE65pt_0QnIageBLEwv2MG2BW-b5NwoM/edit?usp=sharing")

update_strains_W(file = "strains/W.csv",
  from = "https://docs.google.com/spreadsheets/d/1sDqH44-LWbf4_loE_BVDxsehEoiZ44E93apYCtpgs1M/edit?usp=sharing")
}
\arguments{
\item{file}{Where to save the CSV.}

\item{from}{Where to download the CSV. (Must be an accessible google spreadsheet URL)}
}
\description{
Update a CSV file from a google spreadsheet
}
