% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/update.R
\name{update_genes}
\alias{update_genes}
\title{Update gene annotation table}
\usage{
update_genes(to)
}
\arguments{
\item{to}{Where to write the CSV}
}
\description{
Update \emph{S. cerevisiae} gene annotations from
\href{http://www.yeastgenome.org}{SGD}.
}

