% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_genetic_distances}
\alias{update_genetic_distances}
\title{Update genetic distances}
\usage{
update_genetic_distances(to)
}
\arguments{
\item{to}{Where to write the CSV}
}
\description{
Update \emph{S.cerevisiae} genetic distances from 
\href{http://www.yeastgenome.org}{SGD}
}

