% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_genes}
\alias{update_genes}
\title{Update gene annotation table}
\usage{
update_genes(file = "genes/sgd.csv",
  from = "http://downloads.yeastgenome.org/curation/chromosomal_feature/SGD_features.tab")
}
\arguments{
\item{file}{Where to write the CSV.}

\item{from}{Where to download gene annotations from.}
}
\description{
Update *S.cerevisiae* gene annotations from 
[SGD](http://www.yeastgenome.org)
}
