% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_genetic_distances}
\alias{update_genetic_distances}
\title{Update genetic distances}
\usage{
update_genetic_distances(dir = "genetic_distances",
  from = "http://downloads.yeastgenome.org/curation/chromosomal_feature/SGD_features.tab")
}
\arguments{
\item{dir}{Where to write the CSVs. Each chromosome is saved in a separate file.}

\item{from}{Where to download genetic distances from.}
}
\description{
Update \emph{S.cerevisiae} genetic distances from
\href{http://www.yeastgenome.org}{SGD}
}
